adding “responsive” table
This is hacky. Since a lot of the values here are loaded from the column’s text directly, i settled on the trade off of adding additional tds, then showing and hiding them in css. I do know that this does create a technically invalid html table, but we aren’t being judged on this page for SEO, and in my mind it was the easiest way to remodel the code here without changing a bunch of stuff that would take 10x as long.